modifications v2.1

+++ app/src/main/java/com/example/market/ui/product/ProductDetails.kt
@@ -1,6 +1,7 @@
 package com.example.market.ui.product
 
 import android.content.ContentValues.TAG
+import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
 import android.util.Log
@@ -10,16 +11,19 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.AdapterView
 import android.widget.ArrayAdapter
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import com.denzcoskun.imageslider.models.SlideModel
 import com.example.market.R
+import com.example.market.auth.AuthActivity
 import com.example.market.data.pojo.Product
 import com.example.market.databinding.FragmentProductDetailsBinding
 import com.example.market.utils.Constants
 import com.example.market.utils.NetworkResult
 import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
 import kotlin.math.roundToInt
 import kotlin.random.Random
 
@@ -28,8 +32,8 @@ class ProductDetails : Fragment() {
 
     private var _binding : FragmentProductDetailsBinding? = null
     private val binding get() = _binding!!
-    private lateinit var sharedPreferences: SharedPreferences
-    private lateinit var editor: SharedPreferences.Editor
+    @Inject
+    lateinit var sharedPreferences: SharedPreferences
     private val viewModel : ProductDetailsViewModel by viewModels()
 
     private val args : ProductDetailsArgs by navArgs()
@@ -46,13 +50,11 @@ class ProductDetails : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        sharedPreferences = requireContext().getSharedPreferences(Constants.SharedPreferences, 0)
-        editor = sharedPreferences.edit()
 
         viewModel.getProduct(args.productId)
-        viewModel.getProducts()
 
         observeProductResponse()
+
 
 
     }
@@ -107,7 +109,7 @@ class ProductDetails : Fragment() {
         imgSlider.setImageList(imageList)
         binding.titleText.text = product.title
         binding.brandText.text = product.vendor
-        binding.descreptionText.text = product.body_html
+        binding.descriptionText.text = product.body_html
         val sizeList = ArrayList<String>()
         for(i in product.options!![0].values.indices){
             sizeList.add(product.options[0].values[i])
@@ -118,7 +120,7 @@ class ProductDetails : Fragment() {
         }
         setColorList(colorList)
         setSizeList(sizeList,product)
-        binding.priceText.text = product.variants!![0].price
+        binding.priceText.text = product.variants!![0].price +" "+ sharedPreferences.getString(Constants.CURRENCY_KEY,"EGP")
         binding.ratingBar.rating = randomRounded.toFloat()
         checkFavorite(product)
 
@@ -128,15 +130,19 @@ class ProductDetails : Fragment() {
         if (product.isFavourite){
             binding.addToFavoriteButton.setImageDrawable(binding.root.context.getDrawable(R.drawable.ic_filled_heart))
         }
-        binding.addToFavoriteButton.setOnClickListener {
-            if (product.isFavourite) {
-                viewModel.deleteFavourite(product)
-                binding.addToFavoriteButton.setImageDrawable(binding.root.context.getDrawable(R.drawable.ic_heart))
+        binding.addToFavoriteButton.setOnClickListener{
+            if (sharedPreferences.getBoolean(Constants.IS_Logged, false)) {
+                if (product.isFavourite) {
+                    viewModel.deleteFavourite(product)
+                    binding.addToFavoriteButton.setImageDrawable(binding.root.context.getDrawable(R.drawable.ic_heart))
+                } else {
+                    viewModel.addFavourite(product)
+                    binding.addToFavoriteButton.setImageDrawable(binding.root.context.getDrawable(R.drawable.ic_filled_heart))
+                }
+                product.isFavourite = !product.isFavourite
             } else {
-                viewModel.addFavourite(product)
-                binding.addToFavoriteButton.setImageDrawable(binding.root.context.getDrawable(R.drawable.ic_filled_heart))
+                showAlertDialog()
             }
-            product.isFavourite = !product.isFavourite
         }
     }
 
@@ -158,22 +164,24 @@ class ProductDetails : Fragment() {
             }
         }
     }
-    private fun observeProductsResponse() {
-        viewModel.products.observe(viewLifecycleOwner) { response ->
-            when (response) {
-                is NetworkResult.Success -> {
-                    //stopShimmer()
-                    response.data?.let {
-                    }
-                }
-                is NetworkResult.Error -> {
-                    //stopShimmer()
-                }
-                is NetworkResult.Loading -> {
-                    //startShimmer()
-                }
-            }
+    private fun showAlertDialog() {
+        val builder = AlertDialog.Builder(requireContext())
+        builder.setTitle("Error")
+        builder.setMessage("You Must login first.")
+        builder.setIcon(android.R.drawable.ic_dialog_alert)
+        builder.setPositiveButton(resources.getString(R.string.OK)) { _, _ ->
+            val i = Intent(requireActivity(), AuthActivity::class.java)
+            i.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP
+            startActivity(i)
+            activity?.finish()
         }
+        builder.setNegativeButton(resources.getString(R.string.cancel)) { dialogInterface, _ ->
+            dialogInterface.dismiss()
+        }
+
+        val alertDialog: AlertDialog = builder.create()
+        alertDialog.setCancelable(false)
+        alertDialog.show()
     }
 
 }
